requirements.txt
django

djangorestframework
markdown       # Markdown support for the browsable API.
django-filter  # Filtering support

django-braces